+++ x/gitopia/client/cli/txRepository.go
@@ -235,6 +235,128 @@ func CmdRemoveRepositoryCollaborator() *cobra.Command {
 	return cmd
 }
 
+func CmdCreateRepositoryLabel() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "create-repository-label [id] [name] [color] [description]",
+		Short: "Create Repository Label",
+		Args:  cobra.ExactArgs(4),
+		RunE: func(cmd *cobra.Command, args []string) error {
+			id, err := strconv.ParseUint(args[0], 10, 64)
+			if err != nil {
+				return err
+			}
+
+			argsName, err := cast.ToStringE(args[1])
+			if err != nil {
+				return err
+			}
+			argsColor, err := cast.ToStringE(args[2])
+			if err != nil {
+				return err
+			}
+			argsDescription, err := cast.ToStringE(args[3])
+			if err != nil {
+				return err
+			}
+
+			clientCtx, err := client.GetClientTxContext(cmd)
+			if err != nil {
+				return err
+			}
+
+			msg := types.NewMsgCreateRepositoryLabel(clientCtx.GetFromAddress().String(), id, argsName, argsColor, argsDescription)
+			if err := msg.ValidateBasic(); err != nil {
+				return err
+			}
+			return tx.GenerateOrBroadcastTxCLI(clientCtx, cmd.Flags(), msg)
+		},
+	}
+
+	flags.AddTxFlagsToCmd(cmd)
+
+	return cmd
+}
+
+func CmdUpdateRepositoryLabel() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "update-repository-label [repositoryId] [labelId] [name] [color] [description]",
+		Short: "Update Repository Label",
+		Args:  cobra.ExactArgs(5),
+		RunE: func(cmd *cobra.Command, args []string) error {
+			repositoryId, err := strconv.ParseUint(args[0], 10, 64)
+			if err != nil {
+				return err
+			}
+			labelId, err := strconv.ParseUint(args[1], 10, 64)
+			if err != nil {
+				return err
+			}
+
+			argsName, err := cast.ToStringE(args[2])
+			if err != nil {
+				return err
+			}
+			argsColor, err := cast.ToStringE(args[3])
+			if err != nil {
+				return err
+			}
+			argsDescription, err := cast.ToStringE(args[4])
+			if err != nil {
+				return err
+			}
+
+			clientCtx, err := client.GetClientTxContext(cmd)
+			if err != nil {
+				return err
+			}
+
+			msg := types.NewMsgUpdateRepositoryLabel(clientCtx.GetFromAddress().String(), repositoryId, labelId, argsName, argsColor, argsDescription)
+			if err := msg.ValidateBasic(); err != nil {
+				return err
+			}
+			return tx.GenerateOrBroadcastTxCLI(clientCtx, cmd.Flags(), msg)
+		},
+	}
+
+	flags.AddTxFlagsToCmd(cmd)
+
+	return cmd
+}
+
+func CmdDeleteRepositoryLabel() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "delete-repository-label [id] [name]",
+		Short: "Delete Repository Label",
+		Args:  cobra.ExactArgs(2),
+		RunE: func(cmd *cobra.Command, args []string) error {
+			id, err := strconv.ParseUint(args[0], 10, 64)
+			if err != nil {
+				return err
+			}
+
+			argsName, err := cast.ToStringE(args[1])
+			if err != nil {
+				return err
+			}
+
+			clientCtx, err := client.GetClientTxContext(cmd)
+			if err != nil {
+				return err
+			}
+
+			msg := types.NewMsgDeleteRepositoryLabel(clientCtx.GetFromAddress().String(), id, argsName)
+			if err := msg.ValidateBasic(); err != nil {
+				return err
+			}
+			return tx.GenerateOrBroadcastTxCLI(clientCtx, cmd.Flags(), msg)
+		},
+	}
+
+	flags.AddTxFlagsToCmd(cmd)
+
+	return cmd
+}
+
 func CmdCreateBranch() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "create-branch [repo id] [branch name] [commit SHA]",