+++ x/gitopia/types/messages_issue.go
@@ -46,6 +46,21 @@ func (msg *MsgCreateIssue) ValidateBasic() error {
 		return sdkerrors.Wrapf(sdkerrors.ErrInvalidAddress, "invalid creator address (%s)", err)
 	}
 
+	if len(msg.Title) > 255 {
+		return sdkerrors.Wrapf(sdkerrors.ErrInvalidRequest, "title length exceeds limit: 255")
+	} else if len(msg.Title) < 3 {
+		return sdkerrors.Wrapf(sdkerrors.ErrInvalidRequest, "title too short")
+	}
+	if len(msg.Description) > 20000 {
+		return sdkerrors.Wrapf(sdkerrors.ErrInvalidRequest, "description length exceeds limit: 20000")
+	}
+	if len(msg.Assignees) > 10 {
+		return sdkerrors.Wrapf(sdkerrors.ErrInvalidRequest, "can't give more than 10 assignees at a time")
+	}
+	if len(msg.Labels) > 10 {
+		return sdkerrors.Wrapf(sdkerrors.ErrInvalidRequest, "can't give more than 10 labels at a time")
+	}
+
 	if len(msg.Assignees) > 0 {
 		unique := make(map[string]bool, len(msg.Assignees))
 		for _, assignee := range msg.Assignees {
@@ -60,18 +75,29 @@ func (msg *MsgCreateIssue) ValidateBasic() error {
 			}
 		}
 	}
+
+	if len(msg.Labels) > 0 {
+		unique := make(map[string]bool, len(msg.Labels))
+		for _, label := range msg.Labels {
+			if !unique[label] {
+				unique[label] = true
+			} else {
+				return sdkerrors.Wrapf(sdkerrors.ErrInvalidRequest, "duplicate label (%v)", label)
+			}
+		}
+	}
+
 	return nil
 }
 
 var _ sdk.Msg = &MsgUpdateIssue{}
 
-func NewMsgUpdateIssue(creator string, id uint64, title string, description string, labels []string, weight uint64, assignees []string) *MsgUpdateIssue {
+func NewMsgUpdateIssue(creator string, id uint64, title string, description string, weight uint64, assignees []string) *MsgUpdateIssue {
 	return &MsgUpdateIssue{
 		Id:          id,
 		Creator:     creator,
 		Title:       title,
 		Description: description,
-		Labels:      labels,
 		Weight:      weight,
 		Assignees:   assignees,
 	}
@@ -360,6 +386,114 @@ func (msg *MsgRemoveIssueAssignees) ValidateBasic() error {
 			unique[assignee] = true
 		} else {
 			return sdkerrors.Wrapf(sdkerrors.ErrInvalidRequest, "duplicate assignee (%s)", assignee)
+		}
+	}
+	return nil
+}
+
+var _ sdk.Msg = &MsgAddIssueLabels{}
+
+func NewMsgAddIssueLabels(creator string, id uint64, labels []string) *MsgAddIssueLabels {
+	return &MsgAddIssueLabels{
+		Id:      id,
+		Creator: creator,
+		Labels:  labels,
+	}
+}
+
+func (msg *MsgAddIssueLabels) Route() string {
+	return RouterKey
+}
+
+func (msg *MsgAddIssueLabels) Type() string {
+	return "AddIssueLabels"
+}
+
+func (msg *MsgAddIssueLabels) GetSigners() []sdk.AccAddress {
+	creator, err := sdk.AccAddressFromBech32(msg.Creator)
+	if err != nil {
+		panic(err)
+	}
+	return []sdk.AccAddress{creator}
+}
+
+func (msg *MsgAddIssueLabels) GetSignBytes() []byte {
+	bz := ModuleCdc.MustMarshalJSON(msg)
+	return sdk.MustSortJSON(bz)
+}
+
+func (msg *MsgAddIssueLabels) ValidateBasic() error {
+	_, err := sdk.AccAddressFromBech32(msg.Creator)
+	if err != nil {
+		return sdkerrors.Wrapf(sdkerrors.ErrInvalidAddress, "invalid creator address (%s)", err)
+	}
+
+	if len(msg.Labels) < 1 {
+		return sdkerrors.Wrapf(sdkerrors.ErrInvalidRequest, "empty labels list")
+	} else if len(msg.Labels) > 10 {
+		return sdkerrors.Wrapf(sdkerrors.ErrInvalidRequest, "can't give more than 10 labels at a time")
+	}
+
+	unique := make(map[string]bool, len(msg.Labels))
+	for _, label := range msg.Labels {
+		if !unique[label] {
+			unique[label] = true
+		} else {
+			return sdkerrors.Wrapf(sdkerrors.ErrInvalidRequest, "duplicate label (%v)", label)
+		}
+	}
+	return nil
+}
+
+var _ sdk.Msg = &MsgRemoveIssueLabels{}
+
+func NewMsgRemoveIssueLabels(creator string, id uint64, labels []string) *MsgRemoveIssueLabels {
+	return &MsgRemoveIssueLabels{
+		Id:      id,
+		Creator: creator,
+		Labels:  labels,
+	}
+}
+
+func (msg *MsgRemoveIssueLabels) Route() string {
+	return RouterKey
+}
+
+func (msg *MsgRemoveIssueLabels) Type() string {
+	return "AddIssueLabels"
+}
+
+func (msg *MsgRemoveIssueLabels) GetSigners() []sdk.AccAddress {
+	creator, err := sdk.AccAddressFromBech32(msg.Creator)
+	if err != nil {
+		panic(err)
+	}
+	return []sdk.AccAddress{creator}
+}
+
+func (msg *MsgRemoveIssueLabels) GetSignBytes() []byte {
+	bz := ModuleCdc.MustMarshalJSON(msg)
+	return sdk.MustSortJSON(bz)
+}
+
+func (msg *MsgRemoveIssueLabels) ValidateBasic() error {
+	_, err := sdk.AccAddressFromBech32(msg.Creator)
+	if err != nil {
+		return sdkerrors.Wrapf(sdkerrors.ErrInvalidAddress, "invalid creator address (%s)", err)
+	}
+
+	if len(msg.Labels) < 1 {
+		return sdkerrors.Wrapf(sdkerrors.ErrInvalidRequest, "empty labels list")
+	} else if len(msg.Labels) > 50 {
+		return sdkerrors.Wrapf(sdkerrors.ErrInvalidRequest, "can't give more than 50 labels at a time")
+	}
+
+	unique := make(map[string]bool, len(msg.Labels))
+	for _, label := range msg.Labels {
+		if !unique[label] {
+			unique[label] = true
+		} else {
+			return sdkerrors.Wrapf(sdkerrors.ErrInvalidRequest, "duplicate label (%v)", label)
 		}
 	}
 	return nil