+++ x/gitopia/utils/issue.go
@@ -10,12 +10,21 @@ func JoinAssignees(assignees []string) string {
 	res := ""
 	len := len(assignees)
 	for i, a := range assignees {
-		res += " "
-		res += "@"
-		res += a
+		res += fmt.Sprintf(" @%s", a)
 		if i == len-2 && len > 1 {
-			res += " "
-			res += "and"
+			res += " and"
+		}
+	}
+	return res
+}
+
+func JoinLabels(labels []string) string {
+	res := ""
+	len := len(labels)
+	for i, l := range labels {
+		res += fmt.Sprintf(" **%s**", l)
+		if i == len-2 && len > 1 {
+			res += " and"
 		}
 	}
 	return res
@@ -39,6 +48,15 @@ func IssueAssigneeExists(a []string, val string) (int, bool) {
 	return 0, false
 }
 
+func IssueLabelExists(l []uint64, val uint64) (int, bool) {
+	for i, v := range l {
+		if v == val {
+			return i, true
+		}
+	}
+	return 0, false
+}
+
 func IssueUpdateTitleCommentBody(creator string, oldTitle string, newTitle string) string {
 	return fmt.Sprintf("@%v changed title from **~~%v~~** to **%v**", creator, oldTitle, newTitle)
 }
@@ -53,6 +71,14 @@ func IssueAddAssigneesCommentBody(creator string, assignees []string) string {
 
 func IssueRemoveAssigneesCommentBody(creator string, assignees []string) string {
 	return fmt.Sprintf("@%v unassigned"+JoinAssignees(assignees), creator)
+}
+
+func IssueAddLabelsCommentBody(creator string, labels []string) string {
+	return fmt.Sprintf("@%v added"+JoinLabels(labels)+" label", creator)
+}
+
+func IssueRemoveLabelsCommentBody(creator string, labels []string) string {
+	return fmt.Sprintf("@%v removed"+JoinLabels(labels)+" label", creator)
 }
 
 func IssueToggleStateCommentBody(creator string, state types.Issue_State) string {